#42 IT test works

+++ src/test/java/com/jcabi/s3/BucketRule.java
@@ -124,8 +124,10 @@ final class BucketRule implements TestRule {
      */
     private void drop() throws Exception {
         final AmazonS3 aws = this.subj.region().aws();
-        aws.deleteBucket(this.subj.name());
-        Logger.info(this, "S3 bucket %s deleted", this.subj.name());
+        if (aws.doesBucketExist(this.subj.name())) {
+            aws.deleteBucket(this.subj.name());
+            Logger.info(this, "S3 bucket %s deleted", this.subj.name());
+        }
     }
 
 }
